Fix getting setting minimal and maximal scale hint by WmsInstanceLayer, WmsLayerSource, WmsInstanceLayerEntityHandler and MinMax   cherry-picked from c2c3d4cfeb38466d6795f37cf70187081f6045a6
Reverse to old getScaleRecursive-function in WmsLayerSource because of regression bug. Now correct scale and scale hint for sublayer are set
  cherry-picked b07820a51649cf423b7bcced7315f73e3cf87d6f

Fix return  type in getScaleRecursive if no scale is defined but parent is available
If a parent was available the scale of this parent was returned instead of the function call to getScaleRecursive. Other function expexcted a MinMax as return type, also the tree structur could not be resolve if the level was deeper then 2

  cherry-picked 278354053677629343b5f842cbbd489cbd0176f3

Fix getMaxScale & getMinScale if parent scale null
  cherry-picked d25eaab8abe5c8ccaca0c9c7be1cec8bdf75422c

Add missing return in one logic path
  cherry-picked 28405d441e648aabd9ba17bc85eac0b1fe1be5e0