Change visibility of some fields in WmsLayerSource and Wmc Entites
Some fields in both entities had the visibilty set to protected as a workaround for some prehistoric doctrine bug.
But with these visibilities the entities are invalid and doctrine:schema:validate would fail and you could not create
a new db with the console commands. The doctrine bug is now fixed, so we can change the visibilty to the correct protected value.