Change check for this.popup to this.popupWindow
property popup is a function so it always will be trueish. To check if the popup is already open we have to check against
property popupWindow. The old code can cause errors if the legend element is used in a button group and breaks the functionality
of other buttons in this group.