Fix selection frame reset on input rotation change, fix erroneous form submit on enter
Every keyup in the rotation input reseted the selection frame to the center of the screen, because the event argument of the eventlistener callback function was wrongly passed to the _updatedGeometry function
and missinterpreted as the truthy reset parameter. Now the function is wrapped and called with the correct false value for reset.
Pressing enter in one of the input fields of the print form, submitted the form without the necesarry logic which is bound to the print-button. Now we catch all keypress-events and check for enter(Code 13) to prevent this and call t
he correct logic.