ImageTransport: Return null if image download fails
Currently if downloadImage() is unable to download an image, it tries to
do the Gd-Operations on null.

Instead downloadImage() should check if $image is null. With this patch,
if $image is indeed null an error gets logged (since a null image is
undesirable, but a failed download is not necessarily the wrong
behaviour for a system) and returns null. This requires a new argument
to downloadImage() ($logger), which all callers have to pass. Most
already have a $logger that can be used, but LegendHandler.php does not,
so a new normally-null logger is introduced and set from PrintService.